+++ libs/client/cornflow_client/core/instance_solution.py
@@ -3,7 +3,8 @@
 import json
 from abc import ABC, abstractmethod
 from datetime import datetime, timedelta
-from typing import List
+from typing import List, Dict, Union
+import logging as log
 
 from genson import SchemaBuilder
 from jsonschema import Draft7Validator
@@ -431,3 +432,62 @@ class InstanceSolutionCore(ABC):
         """Returns the integer value of the hour (in number) from ts string in format 'YYYY-MM-DDTh:m:s'"""
         datetime_obj = datetime.strptime(string, DATETIME_FORMAT)
         return datetime_obj.hour
+
+
+class CheckCore:
+    def get_check_methods(self) -> list:
+        """
+        Finds all class methods starting with check_ and returns them in a list.
+
+        :return: A list of check methods.
+        """
+        check_methods = [
+            m
+            for m in dir(self)
+            if m.startswith("check_")
+            and callable(getattr(self, m))
+            and m != "check_schema"
+            and m != "check_solution"
+        ]
+        return check_methods
+
+    def launch_all_checks(self) -> Dict[str, Union[List, Dict]]:
+        """
+        Launch every check method and return a dict with the check method name as key
+        and the list/dict of errors / warnings as value.
+
+        It will only return those checks that return a non-empty list/dict.
+        If a check method raises an exception, it will be caught and a generic error
+        message will be added to the checks dictionary.
+        """
+        check_method_names = self.get_check_methods()
+        check_methods = {}
+
+        # Execute each check method and catch any exceptions
+        for method_name in check_method_names:
+            try:
+                check_methods[method_name] = getattr(self, method_name)()
+            except Exception:
+                # If a check fails, add a generic error message
+                check_methods[method_name] = [
+                    {
+                        "error_type": "Check execution error",
+                        "error_message": "The execution of the check has failed, please contact support",
+                    }
+                ]
+                log.warning(
+                    f"The execution of the check {method_name} has failed, please contact support"
+                )
+
+        failed_checks = {}
+        for k, v in check_methods.items():
+            if v is None:
+                continue
+
+            try:
+                if len(v) > 0:
+                    failed_checks[k.split("check_")[1]] = v
+            except TypeError:
+                failed_checks[k.split("check_")[1]] = v
+
+        return dict(failed_checks)