+++ cornflow-server/changelog.rst
@@ -1,3 +1,21 @@
+version 1.2.6
+--------------
+
+- released: 2025-10-31
+- description: bump version code to stay up to date with cornflow-client.
+- changelog:
+    - updated cornflow-client to version 1.2.6
+
+version 1.2.5
+--------------
+
+- released: 2025-10-09
+- description:
+- changelog:
+    - updated cornflow-client to version 1.2.5
+    - fixed OpenID authentication error with Microsoft.
+    - added new fields to return on execution endpoint GET call.
+
 version 1.2.4
 --------------